+++ themes/newspack-block-theme/CHANGELOG.md
@@ -1,3 +1,10 @@
+## newspack-block-theme [1.28.5](https://github.com/Automattic/newspack-workspace/compare/newspack-block-theme@1.28.4...newspack-block-theme@1.28.5) (2026-06-23)
+
+
+### Bug Fixes
+
+* **theme:** stop Custom Fields box overwriting editor post meta ([#375](https://github.com/Automattic/newspack-workspace/issues/375)) ([3738a64](https://github.com/Automattic/newspack-workspace/commit/3738a649164b80508530b75f4cd243b8cf0b94fa))
+
 ## newspack-block-theme [1.28.4](https://github.com/Automattic/newspack-workspace/compare/newspack-block-theme@1.28.3...newspack-block-theme@1.28.4) (2026-06-15)